fix(buy): fix broken buy request

+++ src/app/insurance/effects/advice.ts
@@ -37,9 +37,13 @@ export class AdviceEffects {
       }
     })
     .switchMap((latestAdvice) => {
-      return this.adviceService.removeAdvice(latestAdvice._id)
-        .map((res: Response) => new advice.RemoveLatestInsuranceAdviceSuccess(latestAdvice))
-        .catch(error => Observable.of(new advice.RemoveLatestInsuranceAdviceFailure(error)));
+      if (latestAdvice) {
+        return this.adviceService.removeAdvice(latestAdvice._id)
+          .map((res: Response) => new advice.RemoveLatestInsuranceAdviceSuccess(latestAdvice))
+          .catch(error => Observable.of(new advice.RemoveLatestInsuranceAdviceFailure(error)));
+      } else {
+        return Observable.of({ type: 'NO_ACTION' });
+      }
     });
 
   constructor(private actions$: Actions, private adviceService: AdviceService, private store$: Store<fromInsurance.InsuranceState>) {
